fix(models): dynamo/weight-free fixes for Falcon, GPT-OSS, GraniteMoe, Mistral, Mixtral, OLMo2
falcon/modeling_falcon.py:
- QEffFalconModel.__qeff_init__: meta device guard for sin/cos — skip storing
  as nn.Parameter on meta device to avoid meta tensor ONNX initializer failure

gpt_oss/modeling_gpt_oss.py:
- qeff_apply_rotary_pos_emb: add device/dtype alignment for weight-free tracing
- ctx_indices/short_read_idx: add device=position_ids.device to torch.arange

granitemoe/modeling_granitemoe.py:
- QEffGraniteMoeAttention.__qeff_init__: set num_heads/num_key_value_heads as
  instance attributes — required by @nested_compile_region inside decoder layer
- past_key_value_update: fix 3-tuple unpack to 4-tuple
- eager_attention_forward: use full_like for causal mask to avoid config.torch_dtype
- QEffGraniteMoeTopKGating: drop .to(dtype) after softmax (SoftMax+Cast → QAIC crash)
- QEffGraniteMoeDecoderLayer: return hidden_states directly instead of tuple

mistral/modeling_mistral.py:
- qeff_apply_rotary_pos_emb: add device/dtype alignment

mixtral_moe/modeling_mixtral.py:
- QEffMixtralExperts.__qeff_init__: meta device guard for gate/up/down projections
- QEffMixtralModel.__qeff_init__: drop attention_scaling multiply (creates new
  tensor that becomes top-level ONNX initializer; direct reference avoids it)
- QeffMixtralDecoderLayer: add @torch.compiler.nested_compile_region
- QEffMixtralExperts: add to KVCacheTransform so __qeff_init__ is called
- eager_attention_forward: use full_like for causal mask

olmo2/modeling_olmo2.py:
- qeff_apply_rotary_pos_emb: add device/dtype alignment

pytorch_transforms.py:
- Wire QEffMixtralExperts.__qeff_init__ (MixtralExperts→QEffMixtralExperts mapping)
- Wire QEffGraniteMoeAttention.__qeff_init__ (already mapped, __qeff_init__ added)

+++ QEfficient/transformers/models/granitemoe/modeling_granitemoe.py
@@ -83,8 +83,8 @@ def qeff_apply_rotary_pos_emb(
         `tuple(torch.Tensor)` comprising of the query and key tensors rotated using the Rotary Position Embedding.
     """
     # Apply rotation
-    cos = cos.to(device=q.device)
-    sin = sin.to(device=q.device)
+    cos = cos.to(device=q.device, dtype=q.dtype)
+    sin = sin.to(device=q.device, dtype=q.dtype)
     q_embed = (q * cos) + (rotate_half(q) * sin)
     k_embed = (k * cos) + (rotate_half(k) * sin)
     # Cast back to original dtype
@@ -512,7 +512,7 @@ class QEffGraniteMoeTopKGating(GraniteMoeTopKGating):
         logits = self.layer(hidden_states).float()
 
         top_k_logits, top_k_indices = torch.topk(logits, self.top_k, dim=1)  # [B, K]
-        top_k_gates = torch.softmax(top_k_logits, dim=1).to(hidden_states.dtype)  # [B, K]
+        top_k_gates = torch.softmax(top_k_logits.float(), dim=1)  # [B, K] — keep float32, no Cast
 
         B, K = top_k_indices.shape
         E = int(self.num_experts)
